Fix rebalance errors
Update total LP supply in account of `_mintFee`.

+++ ethereum_contracts/contracts/interfaces/IWormhole.sol
@@ -94,5 +94,5 @@ interface WormholeTokenBridge {
 interface IWETH is IERC20 {
     function deposit() external payable;
 
-    function withdraw(uint amount) external;
+    function withdraw(uint256 amount) external;
 }
